requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.16
APScheduler==3.9.1
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cisco-gnmi==1.0.15
cryptography==35.0.0
dill==0.3.4
distro==1.6.0
genie==21.10
genie.libs.clean==21.10
genie.libs.conf==21.10
genie.libs.filetransferutils==21.10
genie.libs.health==21.10
genie.libs.ops==21.10
genie.libs.parser==21.10
genie.libs.sdk==21.10
gitdb==4.0.9
GitPython==3.1.24
grpcio==1.47.0
idna==3.3
Jinja2==3.0.2
jsonpickle==2.0.0
junit-xml==1.9
lxml==4.9.1
MarkupSafe==2.0.1
multidict==5.2.0
ncclient==0.6.13
netaddr==0.8.0
paramiko==2.11.0
pathspec==0.9.0
prettytable==2.2.1
protobuf==3.20.1
psutil==5.8.0
pyats==21.10
pyats.aereport==21.10
pyats.aetest==21.10
pyats.async==21.10
pyats.connections==21.10
pyats.datastructures==21.10
pyats.easypy==21.10
pyats.kleenex==21.10
pyats.log==21.10
pyats.reporter==21.10.1
pyats.results==21.10
pyats.tcl==21.10
pyats.topology==21.10
pyats.utils==21.10
pycparser==2.20
pyftpdlib==1.5.6
PyNaCl==1.5.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
tftpy==0.8.0
tqdm==4.62.3
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
unicon==21.10
unicon.plugins==21.10
urllib3==1.26.7
wcwidth==0.2.5
xmltodict==0.12.0
yamllint==1.26.3
yang.connector==22.6.2
yarl==1.7.0
